requirements.txt
dj-database-url==0.4.2
Django==1.11
django-chartjs==1.1
django-dashing==0.3.3
django-filter==1.0.2
django-google-charts==0.1.1
gunicorn==19.7.1
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
